Re-organize build targets, part 3.
Remove last traces of the 2.8 EXTRA_LIBS making dependencies
explicit. Make some separate libs which are exported to the
plugins.
