+++ cmake/OcpnFindGtk.cmake
@@ -1,0 +1,24 @@
+if (TARGET ocpn::gtk)
+  return()
+endif ()
+
+IF(NOT OCPN_FORCE_GTK3)
+  FIND_PACKAGE(GTK2)
+ENDIF()
+
+add_library(_GTK INTERFACE)
+IF(GTK2_FOUND)
+  target_include_directories(_GTK INTERFACE ${GTK2_INCLUDE_DIRS})
+  target_link_libraries(_GTK INTERFACE ${GTK2_LIBRARIES})
+  MESSAGE(STATUS "Building against GTK2...")
+  MESSAGE(STATUS "GTK2 headers: ${GTK2_INCLUDE_DIRS}")
+  MESSAGE(STATUS "GTK2 libs: ${GTK2_LIBRARIES}")
+ELSE(GTK2_FOUND)
+  FIND_PACKAGE(GTK3)
+  target_include_directories(_GTK INTERFACE ${GTK3_INCLUDE_DIRS})
+  target_link_libraries(_GTK INTERFACE ${GTK3_LIBRARIES})
+  target_compile_definitions(_GTK PUBLIC  "-D__WXGTK3__")
+  MESSAGE(STATUS "Building against GTK3...")
+ENDIF(GTK2_FOUND)
+target_link_libraries(${PACKAGE_NAME} PRIVATE _GTK)
+add_library(ocpn::gtk ALIAS _GTK)
