+++ libs/ssl_sha1/src/sha1.c
@@ -35,7 +35,7 @@
 #if defined(POLARSSL_SHA1_C)
 
 //#include "polarssl/sha1.h"
-#include "sha1.h"
+#include "ssl/sha1.h"
 
 #if defined(POLARSSL_FS_IO) || defined(POLARSSL_SELF_TEST)
 #include <stdio.h>
