+++ libs/tinyxml/src/tinyxml.cpp
