+++ libs/tinyxml/src/tinyxmlerror.cpp
