+++ libs/tinyxml/src/tinyxmlparser.cpp
