+++ plugins/grib_pi/CMakeLists.txt
@@ -27,6 +27,10 @@ IF (COMMAND cmake_policy)
   endif (POLICY CMP0043)
 ENDIF (COMMAND cmake_policy)
 
+if (NOT CMAKE_MODULE_PATH)
+  set (CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../../cmake)
+endif ()
+
 PROJECT(grib_pi)
 
 SET(PACKAGE_NAME grib_pi)
@@ -162,6 +166,9 @@ ADD_DEFINITIONS(-DJASPER)
 ADD_DEFINITIONS(-DEXCLUDE_MIF_SUPPORT -DEXCLUDE_PNM_SUPPORT
                 -DEXCLUDE_BMP_SUPPORT -DEXCLUDE_RAS_SUPPORT
                 -DEXCLUDE_JPG_SUPPORT -DEXCLUDE_PGX_SUPPORT)
+
+
+
    
 add_subdirectory("libs/jasper")
 INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR}/libs/jasper/src/include)
@@ -223,21 +230,26 @@ IF(WIN32)
   ADD_DEPENDENCIES(${PACKAGE_NAME} ${PARENT})
 ENDIF(WIN32)
 
-TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} ${wxWidgets_LIBRARIES} )
-TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} ${OCPN_Core_LIBRARIES} )
-TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} WXJSON )
-TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} JASPER )
+TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE ${wxWidgets_LIBRARIES} )
+TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE ${OCPN_Core_LIBRARIES} )
+TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE WXJSON )
+TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE JASPER )
 
-target_link_libraries(${PACKAGE_NAME} wxsvg::wxsvg)
+target_link_libraries(${PACKAGE_NAME} PRIVATE wxsvg::wxsvg)
 
 IF(NOT WIN32)
-TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} "-lz" )
+    TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE "-lz" )
 ENDIF()
 
 IF(APPLE)
  FIND_PACKAGE(ZLIB REQUIRED)
- TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} ${ZLIB_LIBRARIES} )
+ TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE ${ZLIB_LIBRARIES} )
 ENDIF(APPLE)
+
+if  (NOT WIN32 AND NOT APPLE AND NOT QT_ANDROID)
+  include(OcpnFindGtk)
+  target_link_libraries(${PACKAGE_NAME} PRIVATE ocpn::gtk)
+endif ()
 
 IF(UNIX AND NOT APPLE AND NOT QT_ANDROID)
     FIND_PACKAGE(BZip2 REQUIRED)
@@ -245,7 +257,9 @@ IF(UNIX AND NOT APPLE AND NOT QT_ANDROID)
     FIND_PACKAGE(ZLIB REQUIRED)
     INCLUDE_DIRECTORIES(${ZLIB_INCLUDE_DIR})
 
-    TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} ${BZIP2_LIBRARIES} ${ZLIB_LIBRARY})
+    TARGET_LINK_LIBRARIES( ${PACKAGE_NAME} PRIVATE
+        ${BZIP2_LIBRARIES} ${ZLIB_LIBRARY}
+    )
 ENDIF(UNIX AND NOT APPLE AND NOT QT_ANDROID)
 
 IF( APPLE )
