+++ plugins/wmm_pi/CMakeLists.txt
@@ -30,7 +30,7 @@ IF( NOT CMAKE_BUILD_TYPE )
        FORCE )
 ENDIF( NOT CMAKE_BUILD_TYPE )
 
-
+INCLUDE_DIRECTORIES(${CMAKE_SOURCE_DIR}/include ${CMAKE_SOURCE_DIR}/src)
 
 MESSAGE (STATUS "Build type: ${CMAKE_BUILD_TYPE}")
