+++ src/iENCToolbar.cpp
@@ -23,6 +23,8 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301,  USA.         *
  **************************************************************************/
 
+#include "config.h"
+
 #include "wx/wxprec.h"
 
 #ifndef  WX_PRECOMP
