+++ src/undo.cpp
@@ -26,6 +26,8 @@
  *
  */
 
+#include "config.h"
+
 #include "wx/wxprec.h"
 
 #ifndef  WX_PRECOMP
